+++ lib/devbar_plugins/log_network.dart
@@ -1,4 +1,7 @@
 export '../src/devbar/plugins/log_network/http_client.dart'
     show DevbarHttpClient;
 export '../src/devbar/plugins/log_network/plugin.dart'
-    show LogNetworkPlugin, LogNetworkPluginDevbarExtension;
+    show
+        LogNetworkPlugin,
+        LogNetworkPluginDevbarExtension,
+        devbarHttpClientSource;