+++ lib/devbar_plugins/log_queries.dart
@@ -1,0 +1,2 @@
+export '../src/devbar/plugins/log_queries/plugin.dart'
+    show LogQueriesPlugin, LoggedQuery, LogQueriesPluginDevbarExtension;